:sparkles: Adds Markdown
And adjusts colors
