+++ package.json
@@ -25,23 +25,23 @@
     "test.run": "vitest --run"
   },
   "devDependencies": {
-    "@builder.io/qwik": "1.2.6",
-    "@builder.io/qwik-city": "1.2.6",
+    "@builder.io/qwik": "1.2.7",
+    "@builder.io/qwik-city": "1.2.7",
     "@tailwindcss/forms": "0.5.4",
     "@tailwindcss/typography": "0.5.9",
     "@trivago/prettier-plugin-sort-imports": "4.2.0",
     "@types/eslint": "8.44.2",
-    "@types/node": "20.4.10",
-    "@typescript-eslint/eslint-plugin": "6.3.0",
-    "@typescript-eslint/parser": "6.3.0",
-    "autoprefixer": "10.4.14",
+    "@types/node": "20.5.1",
+    "@typescript-eslint/eslint-plugin": "6.4.0",
+    "@typescript-eslint/parser": "6.4.0",
+    "autoprefixer": "10.4.15",
     "eslint": "8.47.0",
-    "eslint-plugin-qwik": "1.2.6",
+    "eslint-plugin-qwik": "1.2.7",
     "husky": "8.0.3",
-    "lint-staged": "13.2.3",
-    "postcss": "8.4.27",
-    "prettier": "3.0.1",
-    "prettier-plugin-tailwindcss": "0.5.2",
+    "lint-staged": "14.0.0",
+    "postcss": "8.4.28",
+    "prettier": "3.0.2",
+    "prettier-plugin-tailwindcss": "0.5.3",
     "tailwind-merge": "1.14.0",
     "tailwindcss": "3.3.3",
     "tailwindcss-hero-patterns": "0.1.2",
@@ -49,7 +49,7 @@
     "undici": "5.23.0",
     "vite": "4.4.9",
     "vite-tsconfig-paths": "4.2.0",
-    "vitest": "0.34.1"
+    "vitest": "0.34.2"
   },
   "lint-staged": {
     "!(/)**/*.{ts,tsx}": [
