+++ src/content/info.mdx
@@ -1,0 +1,12 @@
+import { inLocaleTime, inRelativeTime } from '~/utils';
+
+## Where
+
+- Venue: **Smith Hanok**
+- City: **Seoul**
+- Maps: **[Google](https://goo.gl/maps/jueppkxY3FdjdCFu7)**
+
+## When
+
+- Time: <time dateTime={new Date('April 12 2024 11:30 GMT+9').toISOString()}>**11:30 AM - 2:30 PM**</time>
+- Date: <time dateTime={new Date('April 12 2024 11:30 GMT+9').toISOString()}>**April 12th, 2024**</time>
